Move icon-related helper structs into bevy_render::texture::icon
